fix type check error

+++ service/frontend/workflowHandler.go
@@ -165,7 +165,7 @@ func NewWorkflowHandler(
 	return &WorkflowHandler{
 		Resource:        resource,
 		config:          config,
-		healthStatus:    HealthStatusOK,
+		healthStatus:    int32(HealthStatusOK),
 		tokenSerializer: common.NewJSONTaskTokenSerializer(),
 		rateLimiter: quotas.NewMultiStageRateLimiter(
 			func() float64 {
